Lacayo - Improve initial setting of wave
Now the wave is formed first by rotating the Up direction to be
facing the target planet, and the initial position is set away
from the origin by a given distance.

Wave will now need a reference to the player and the target
planet. As such, created an empty target planet for temp use
in LacayoScene.

Refactored methods.