Lacayo - Add difference between wave edge and space
Wave edge is marked by a black rectangle that is the trigger collider
for the wave.

Wave space is the empty "space" that is supposed to be swallowed up
by the wave.

Added both sprites to a new render layer called Wave that is
on top of previously made layers.