Lacayo - Add collision between player and wave
For some reason Wave is not able to get the BoxCollider2D
component from its first child, so a hack was to simply serialize
the child and set it in the inspector so it calls
`edge.GetComponent<BoxCollider2D>();` which works but is not the
best solution.

In any case, for the method to be complete, need a method from
Player to end the game (added TODO as result).