Lacayo - Add public method to check if planet is behind wave
Using the same dot product method with getting the lateral
distance the wave needs to keep in line with the player, this
method will return the distance a given position of a planet is
behind the edge of the wave.

This allows the wave to check, by calling this method, if it's far
enough behind the wave that it should be despawned.