Lacayo - Add parallax to stars
Background.cs handles the children stars that would be attached
to the Background game object in the scene.

Firstly, Start grabs the transform of each star and pairs it with
a parallax coefficient that is the percentage the star will move
with the camera.

In Update, it iterates over all stars and updates their positions
based on the camera movement. If a star were to leave the frame,
it will spawn on the other side of that frame.